APR-159 moved add roatp organisation models into separate class files
