+++ src/SFA.DAS.AssessorService.Api.Types/Models/Roatp/OrganisationStatus.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SFA.DAS.AssessorService.Api.Types.Models.Roatp
+{
+    public class OrganisationStatus
+    {
+        public int Id { get; set; }
+        public string Status { get; set; }
+    }
+}
