+++ src/SFA.DAS.AssessorService.Api.Types/Models/Roatp/ProviderType.cs
@@ -1,0 +1,10 @@
+﻿namespace SFA.DAS.AssessorService.Api.Types.Models.Roatp
+{
+    
+    public class ProviderType
+    {
+        public int Id { get; set; }
+        public string Type { get; set; }
+        public string Description { get; set; }
+    }
+}
