+++ src/SFA.DAS.AssessorService.Api.Types/Models/Roatp/RemovedReason.cs
@@ -1,0 +1,9 @@
+﻿namespace SFA.DAS.AssessorService.Api.Types.Models.Roatp
+{
+    public class RemovedReason
+    {
+        public int Id { get; set; }
+        public string Reason { get; set; }
+        public string Description { get; set; }
+    }
+}
